+++ config/routes.rb
@@ -1,4 +1,4 @@
 Konacha::Engine.routes.draw do
-  match "/"      => "konacha/specs#index"
-  match "/*spec" => "konacha/specs#show"
+  match "/"     => "konacha/specs#specs"
+  match "*path" => "konacha/specs#specs"
 end