+++ spec/models/spec_spec.rb
@@ -31,16 +31,27 @@ describe Konacha::Spec do
   end
 
   describe ".find" do
-    it "returns the Spec with the given asset_name" do
-      all = [described_class.new("a_spec.js"),
-             described_class.new("b_spec.js")]
-      described_class.should_receive(:all) { all }
-      described_class.find("b_spec").should == all[1]
+    it "returns all Specs if given an empty path" do
+      all = ["a_spec.js", "b_spec.js"]
+      Konacha.should_receive(:spec_paths) { all }
+      described_class.find("").map(&:path).should == all
     end
 
-    it "returns nil if no such spec exists" do
-      described_class.should_receive(:all) { [] }
-      described_class.find("b_spec").should be_nil
+    it "returns an array containing the Spec with the given asset_name" do
+      all = ["a_spec.js", "b_spec.js"]
+      Konacha.should_receive(:spec_paths) { all }
+      described_class.find("b_spec").map(&:path).should == [all[1]]
+    end
+
+    it "returns Specs that are children of the given path" do
+      all = ["a/a_spec_1.js", "a/a_spec_2.js", "b/b_spec.js"]
+      Konacha.should_receive(:spec_paths) { all }
+      described_class.find("a").map(&:path).should == all[0..1]
+    end
+
+    it "raises NotFound if no Specs match" do
+      Konacha.should_receive(:spec_paths) { [] }
+      expect { described_class.find("b_spec") }.to raise_error(Konacha::Spec::NotFound)
     end
   end
 end