+++ spec/views/specs/show.html.erb_spec.rb
@@ -1,9 +1,0 @@
-require 'spec_helper'
-
-describe "konacha/specs/show" do
-  it "renders a script tag for @spec" do
-    assign(:spec, spec_double("spec"))
-    render
-    rendered.should have_selector("script[src='/assets/spec.js?body=1']")
-  end
-end