+++ History.md
@@ -5,6 +5,7 @@
 * Switch default port to 3500 to avoid collisions with HTTP proxies (like Charles)
 * Update vendored copies of mocha (0.14.0+) and chai (0.4.2+)
 * Do not require spec dependencies multiple times (#3)
+* Support running a subset of specs via a subdirectory path
 
 # 0.9.0